Extract Mixpanel service tests into own file
